Validações e tratamento de erros finalizadas

+++ api/controllers/AbrigosController.js
@@ -4,6 +4,9 @@ class AbrigosController{
     static async PegaTodosAbrigos (req,res){
         try {
             const todosAbrigos = await database.Abrigos.findAll();
+            if (todosAbrigos.length <= 0){
+                return res.status(200).json("Não há abrigos registrados");
+            }
             return res.status(200).json(todosAbrigos);
         } catch (error) {
             return res.status(500).json(error.message);
@@ -14,6 +17,9 @@ class AbrigosController{
         try {
             const { id } = req.params;
             const umAbrigo = await database.Abrigos.findOne({where: {id:Number(id)} });
+            if (umAbrigo == null){
+                return res.status(200).json("Não há abrigo registrado com esse id");
+            }
             return res.status(200).json(umAbrigo);
         } catch (error) {
             return res.status(500).json(error.message);
@@ -34,8 +40,11 @@ class AbrigosController{
         try {
             const { id } = req.params;
             const abrigo = req.body;
-            await database.Abrigos.update(abrigo,{ where: {id:Number(id)} });
             const umAbrigo = await database.Abrigos.findOne({where: {id:Number(id)} });
+            if (umAbrigo == null){
+                return res.status(200).json("Não há abrigo registrado com esse id");
+            }
+            await database.Abrigos.update(abrigo,{ where: {id:Number(id)} });
             return res.status(200).json(umAbrigo);
         } catch (error) {
             return res.status(500).json(error.message);
@@ -45,6 +54,10 @@ class AbrigosController{
     static async DeletaUmAbrigo (req, res){
         try {
             const { id } = req.params;
+            const umAbrigo = await database.Abrigos.findOne({where: {id:Number(id)} });
+            if (umAbrigo == null){
+                return res.status(200).json("Não há abrigo registrado com esse id");
+            }
             await database.Abrigos.destroy({where: {id:Number(id)} });
             return res.status(200).json({message:`O abrigo com o id ${id} foi deletado`});
         } catch (error) {
