+++ api/controllers/TutoresController.js
@@ -4,6 +4,9 @@ class TutoresController{
     static async PegaTodosTutores (req,res){
         try {
             const todosTutores = await database.Tutores.findAll();
+            if (todosTutores.length <= 0){
+                return res.status(200).json("Não há tutores registrados");
+            }
             return res.status(200).json(todosTutores);
         } catch (error) {
             return res.status(500).json(error.message);
@@ -14,6 +17,9 @@ class TutoresController{
         try {
             const { id } = req.params;
             const umtutor = await database.Tutores.findOne({where: {id:Number(id)} });
+            if (umtutor == null){
+                return res.status(200).json("Não há um tutor registrado com esse id");
+            }
             return res.status(200).json(umtutor);
         } catch (error) {
             return res.status(500).json(error.message);
@@ -34,8 +40,11 @@ class TutoresController{
         try {
             const { id } = req.params;
             const tutor = req.body;
-            await database.Tutores.update(tutor,{ where: {id:Number(id)} });
             const umtutor = await database.Tutores.findOne({where: {id:Number(id)} });
+            if (umtutor == null){
+                return res.status(200).json("Não há um tutor registrado com esse id");
+            }
+            await database.Tutores.update(tutor,{ where: {id:Number(id)} });
             return res.status(200).json(umtutor);
         } catch (error) {
             return res.status(500).json(error.message);
@@ -45,6 +54,10 @@ class TutoresController{
     static async DeletaUmTutor (req, res){
         try {
             const { id } = req.params;
+            const umtutor = await database.Tutores.findOne({where: {id:Number(id)} });
+            if (umtutor == null){
+                return res.status(200).json("Não há um tutor registrado com esse id");
+            }
             await database.Tutores.destroy({where: {id:Number(id)} });
             return res.status(200).json({message:`O tutor com o id ${id} foi deletado`});
         } catch (error) {
